services/project-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]